feat!: Add a DatadogLogger (#5)
This adds a Logger that is backed by Datadog.

It does not really belong in reps-do but we are keeping it here for now.

I got to what I have by a combination of reverse engineering
@datadog/browser-logs as well as looking at the incomplete docs at
https://docs.datadoghq.com/api/latest/logs/#send-logs

The Logger sends the pending messages to datadog's http endpoint every 10s by
default.

The logger support an abort signal that aborts the timers etc. The intended
usage is to abort the timers when the server closes the socket.