feat: Add authentications and authorization support to reps-do (#15)
Adds support for customer's providing an `AuthHandler`.  When a connection request is received this 
`AuthHandler` is invoked with the `auth` value and `roomID` from the connection request.   The 
`AuthHandler` should validate that the user authenticated by `auth` is authorized to access the 
room with `roomID`, and return a promise which resolves to `UserData` for the user if authentication
 and authorization is successful, or reject if authentication or authorization fail.
 
 See larger design at: https://www.notion.so/replicache/WebSocket-Authentication-Authorization-and-Security-bf2b1a2e31844efca73c4aa453fdecdb