Fix: Recover from disconnects more robustly. (#5)
Problem: I was assuming that when you call socket.close(), the `close`
event fires immediately but this isn't always true. When it fires
asynchronously the code could get into a state where `Client` thought
the connection was open but the socket thought it was closed. So
we'd never recover because the app would keep heartbeating as if it
was open, but the heartbeat ping would fail because socket was actually
closed.

Solution: Call _disconnect() directly rather than relying on `close`
event to fire and call it.