fix: serialize pokes (#6)
Problem: Ocassionally we would see unexpected "out of order poke"
error messages in the client, especially when a user is using the
app concurrently with a reload together with heavy load. Manual
investigation shows that this can happen if a second poke comes
in before first is done applying. Could not reproduce locally but
simulated by slowing down poke application artifically.

Solution: Lift the Lock class used in Replicache and reps-do to
serialize handlePoke().