fix!: Take 2 of adding Logger support to the worker (#35)
The original configuration approach in 17cd534544fbcf8f2af463ee4ed71abb8ff9f1ba does not work, 
because environment variables (such as a datadog api key, etc) are not available when creating the worker.  They are available when fetch is called.

Also the DatadogLogger needs to be flushed, and `waitUntil` needs to be used on the resulting promise so that this flush can complete after the response has been returned from the worker (see Cloudflare docs on [using Logging services](https://developers.cloudflare.com/workers/learning/logging-workers#set-up-a-logging-service)).

Question: Should we cache the result of createLogger?   Answer: Doesn't seem necessary because a Logger is lightweight.

BREAKING CHANGE: `WorkerOptions` have changed in a breaking manner.  `createLogger` and `getLogLevel` must be used instead of `logger` and `logLevel`.