fix: Fix race condition in handleClose resulting in 'client not found' errors (#43)
### Problem
If a client tries to reconnect, while the previous connection is still open on the server, a race condition occurs and handleClose deletes the new connections entry in the client map (when it intended to delete the entry for the previous connection)

### Solution
Update `handleClose` to only delete the `ClientMap` entry for `clientID` if the entry's WebSocket is the WebSocket whose "close" event is being handled.  If it is not equal, this indicates there is a new `ClientMap` entry for `clientID` from a new connection, and it should not be deleted.