fix: Close WebSocket connection if client entry is not found for push. (#44)
### Problem
If when processing a push message the ClientMap entry for a client is not found, no future push messages over that WebSocket connection will succeed and the client will be wedged.   This should never occur, but may due to bugs such as the one fixed in: 137f39eade69dced38a30c1ae1f2e8e7b36c12ea

### Solution
To be more robust against future bugs, unwedge the client by closing the connection.  The client can then reconnect and create a new ClientMap entry.  Log an error message since this is truly unexpected and should be investigated if it occurs.