fix: Add clientID to pushBodySchema (#78)
clientID is included in push body by replicache but no in the server schema and so push message validation currently fails on the server. 

A follow up would be to update all uses of `s.object` to `s.type` to allow for unknown fields.  It is also believed this would perform better (but should be verified): https://docs.superstructjs.org/resources/faq#how-can-i-allow-unknown-keys-with-object-structs