fix: Revert to replicache@9.0.0-beta.0 (pre-mutation recovery) and update version to 0.3.1  (#13)
Replicache mutation recovery does not work with the reflect's socket based adaptation of the Replicache
stateless protocol.  

Reflect assumes all messages over a socket are for the same client id (the client id that opened the connection).
With mutation recovery a client will push mutations that belong to a different client id.

None of our customers have deployed reflect with replicache@9.0.0 to production, so it is safe to 
revert back to this older version without thinking through the persisted data format changes.