feat: Garbage collect AuthDO connections state, update version to 0.8.0 (#99)
### Problem
Currently the AuthDO's `connection/` state grows in an unbounded way.  

### Solution
Use a Cloudflare Worker Cron trigger to periodically GC the AuthDO's `connection/` state.  

When the worker receives a schedule cron trigger it sends a `/api/auth/v0/revalidateConnections` request to the AuthDO.  The AuthDO then gathers a list of all `roomID`s for which it has `ConnectionRecord`s in it `connection/` state.  It sends the RoomDO for each of these `roomId`s a `/api/auth/v0/connections` request.  The RoomDO responds with a list of it's currently open socket connections.  The AuthDO then deletes any `ConnectionRecord`s in it `connection/` state which are not included in these list of currently open socket connections. 

The new `revalidateConnections` and `connections` auth api endpoints are currently only internal and not exposed by the worker to customers.
   
In the future `revalidateConnections` can be extended to implement "auth revalidation".  We would add to each `ConnectionRecord` a `lastAuthTimestamp` and the `authToken` for the connection.  During `revalidateConnections` we would call the customer provided `authHandler` with `authToken` for any connection whose `lastAuthTimestamp` is more than 10 minutes old.  If the `authHandler` rejects, we tell the `RoomDO` to close that connection.