feat: Add helper functions for pending with makeScanResult (#11)
`makeScanResult` takes an `AsyncIterable`  but sometimes the backend
might use different data store without a unified entry point. For these
cases you might want to merge a "pending" iterator on top of a
"persisted" iterator. To do this we provide a `mergeAsyncIterators`
function as well as a `filterAsyncIterator` function. The filter version
is useful if you want to filter out tombstones from the iterator.