fix!: Make AsyncIterableIteratorToArray an interface (#40)
Previously we had a concrete class called
`AsyncIterableIteratorToArrayWrapper`. When you have a concrete class it
leads to problems because `ReadTransaction` and `WriteTransaction` now
depend on a concrete class. When you have a concrete class you can not
use duck typing. You can not reuse the `ReadTransaction` and
`WriteTransaction` on the client and server unless they point at the same
exact file defining the concrete type.

By making this an interface it is OK to have duplicate compatible
definition of the type.

This removes the possibility to reuse the old
`AsyncIterableIteratorToArrayWrapper` class in code that depends on
replicache.

BREAKING CHANGE!