refactor: Introduce InternalValue (#136)
InternalValue represents a readonly JSON value that we use internally in
Replicache. It is an opaque type that is not exposed to the user.
Conceptually, all the API endpoints use `toInternalValue` and
`fromInternalValue` which clones the json to ensure no mutations of
internal value can ever occur. However, for performance reasons we do
not always clone. In those cases we use `safeCastToJSON` which does no
copying.

To make sure we never copy a value twice, we keep track of the internal
values in a `WeakSet`.

To make things clearer where these conversions are used, these methods
takes a "reason" enumeration.

There is also a config flag that disables the internal value assertions
which also allows skipping creating the WeakSet.

Towards #56