fix: IDBDatabasesStore test isolation. (#191)
Problem
=======
We occasionally see the following error in tests
```
      An error was thrown in a Promise outside a test. Did you forget to await a function or assertion?
      InvalidStateError: Failed to execute 'transaction' on 'IDBDatabase': The database connection is closing.
        at readImpl (src/kv/idb-store.ts:158:16)
        at IDBStore.withRead (src/kv/idb-store.ts:35:17)
```
Solution
======
This is because of a bug in the test isolation logic for IDBDatabasesStore.  We were not actually
namespacing the IDBDatbasesStore's IndexedDB in tests.   

Due to this the collect-idb-databases.test.ts, was deleting IndexedDB databases being used by 
other tests at the same time (due to parallel test execution).   

Fix the logic so we actually namespace the IDBDatbasesStore's IndexedDB in tests.