fix: update test to close idb stores before trying to delete the dbs. (#192)
The test updated here currently work because in src/kv/idb-store.ts we currently close idb databases when we get a onversionchange event indicating someone is trying to delete the idb database.  

```
function openDatabase(name: string): Promise<IDBDatabase> {
  const req = indexedDB.open(name);
  req.onupgradeneeded = () => {
    const db = req.result;
    db.createObjectStore(OBJECT_STORE);
  };
  const wrapped = wrap(req);
  void wrapped.then(db => {
    // *** this is the relevant auto close when someone tries to delete code ***
    db.onversionchange = () => db.close();
  });
  return wrapped;
}
```
Otherwise they would hang (because we would be awaiting a delete that would never complete). 

I don't want these test to rely on the auto-closing behavior, because it seems like something we may want to change 
in the future.

Updated the structure of the test to instead close the idb stores before try to delete the dbs.