feat: Add support for DisconnectHandler (#130)
Customers have requested support for modifying room state in response to a 
client disconnecting from the room.  

This change adds this support via a new `disconnectHandler?: DisconnectHandler` option in `ReflectServerOptions`.

A `DisconnectHandler` can modify room state in response to a client disconnecting 
from the room.  These changes will be synced to all clients of the room just like mutator 
changes. When a client disconnects the `DisconnectHandler` will be called and passed 
a `write: WriteTransaction` for modifying room state.  `write.clientID` will be the id of the
disconnected client.