refactor(DD31): Add put commit variants to WriteTransactionImpl and db.Write (#251)
For DD31 `persist` needs to be able to rebase multiple `Commit<LocalMeta>`a in a single `dag.WriteTransaction`. 
This refactor enables this.  

This refactor also fixes an issue where `refresh` was updating the `sync` head when it should have been updating the `refresh` head, by making `WriteTransactionImpl.commit` take a head name arg rather than inferring it from whether or not the transaction is a rebase.