fix: Precompute the size of B+Tree node entries (#273)
We now compute the size of the B+Tree nodes up front and cache that
result in the entry. The entry used to be a pair, now it is a 3 element
tuple. We only do this for BTreeWrite because for BTreeRead we do not
need to know the size.

This does not change the chunk format. For the chunk we still use a pair
with the key and the value.

Towards #267