fix: initClientDD31 should only look at branches
We were looping over all clients to find the best branch to "fork" from
when we could just loop over the branches directly.

Looping over the branches directly is better because it allows us to
reuse good branches even if there are no clients left referencing them.

This required us to fix a test which was now reusing a branch that it
previously did not consider a candidate.

Towards #165