chore: Remove updateClient (#296)
updateClient was needed when we had a n async (non microtask) hash
computation function in the persist phase. The code was read the current
hash, compute the new hash and then do a write transaction. If the
current hash still matched it then wrote the new chunks with the
computed hashes. If it didn't match we tried again (a few times). Now
that we use an UUID instead of computing the hash we can remove this
read and write retrial logic.

Towards #165