Problem: Vercel comment links break on our docs site when pointing to a different page.
Solution:

This is caused because Vercel is using the <link rel="canonical">
tag and ours has a path component with a double-slash:
<link rel="canonical"
  href="https://doc.replicache.dev//app-structure" data-rh="true">.

This is allowed by the spec, but not our intent. Everything else in
our stack ignores the double-slash so we don't see any bugs
elsewhere.

The canonical tag is generated by our docusaurus "url" configuration
which had a trailing slash.

The docusaurus docs say it shouldn't have that:

https://docusaurus.io/docs/api/docusaurus-config#url