feat(dd31): Mutation Recovery (#331)
* feat(dd31): Mutation recovery

For DD31 we recover mutations from different branches (aka client
groups). We also recover old SDD mutations. The idb databases store
contains the replicache format version and this is used to branch the
mutation recovery code. This means that we have to keep around both
pullSDD and pushSDD as long as we want to support recovering SDD
mutations.

This code has a lot of code duplication and we really need to go back
and dedupe this code now that it turns out that we do not to keep around
the SDD code paths for a while.

The test for mutation recovery runs the SDD tests both when DD31 is true
and false. The old tests where updated to create DD31 dags and then
finally a few selected tests were created to test out recovering both
SDD and DD31 dags at the same time.

Towards #165