refactor: rework refresh to minimize cache misses with memdag locked (#344)
Problem
======
The current refresh implementation will commonly block mutations, pulls and queries on idb reads.  This will often cause jank.  This is because the current implementation simply sets the head in the memdag and then uses the LazyStore's lazy-cache-loading to pull the perdag chunks into the memdag *while the memdag is locked*.

Solution
======
Without a lock on the memdag, gather not-cached chunks from the perdag.  Then lock the memdag, put the gathered chunks, rebase mutations and diff. 

To avoid pulling the entire perdag graph into the memdag the amount of chunk data gathered is limited by size (this change sets the limit to 5MB).  In steady state the diffs gathered in refresh will be << 5MB.

Chunks are gathered in the following order:
1. commits in order from newest to oldest
2. within a commit value btree before index btrees
3. value btree and index btree are visited greedily (kind of a hi-bred of BFS and DFs), we start reading all of a node's child nodes in parallel, when a node loads we gather it, and recurse (i.e. start reading all of the node's child nodes in parallel).  index btrees are gathered in parallel.

 
A write lock is not held on the memdag while not-cached chunks are gathered.  This means that the `isCached` value for chunks can change during the gather process due to cache population, cache eviction and/or chunk GC in the memdag.  So when the write lock is acquired on the memdag below gatheredChunks may contain some chunks that are already cached and may be missing some chunks that are not-cached.  It may also be missing chunks that are not-cached due to the gather size limit.  This is OK, because the already cached chunks will just be re-put into the memdag (a no-op), and the missing not-cache chunks will just be cache misses that will then be loaded from the perdag and put into the cache.  This gather and write approach aims to minimize cache misses during refresh's rebase and diff steps.  This is important because cache misses are relatively slow (as they require reading from idb), and thus handling them while having the memdag locked often creates jank by blocking local mutations, pulls and queries on idb.  This approach does not eliminate all cache misses, but it 
does minimize them.