feat!: Use deep frozen values in stores
BREAKING CHANGE!

Bottom up:

`kv.Store` get now returns a deeply frozen value.
- IDBStore deeply freezes the result of get.

Top down:
- `ReadTransaction` and `WriteTransaction` `get`/`scan` now return a
  deeply frozen value(s).
- `WriteTransaction` `put` now deeply freezes the value.
- We deeply freeze the args passed to a mutator.

We assert that all chunk data is frozen. Even, `clients` and
`clientGroups` chunks. This could have been skipped but things gets
easier to reason about if all chunk data is frozen.

This is all conditional and the freezing  asserts for frozen values are
only done in debug mode.

`kv.Store` still uses the `ReadonlyJSONValue` because that is the
interface `ExperimentalKVStore` uses and we do not want to the
`FrozenJSONValue` to be part of the public API.

Closes #77
Closes #194
Closes rocicorp/replicache#885