doc: Correctly handle SQL exceptions inside mutations.
As pointed out in a discord thread, when a SQL exceptions happens,
the containing transaction is toast and cannot do any more work.
So the way we were trying to skip mutation errors didn't work if the
exception originated in SQL.

Fix this issue by allowing any exception from inside a mutation to
rollback the tx, then starting a new tx to explicitly set the
lastMutationID.

Update comments to match.

[1] https://discord.com/channels/830183651022471199/1042343134786502676