refactor: Add a baseSnapshotHash to LocalMetaDD31 (#417)
Problem
=======
Refresh needs the BaseSnapshot commit of the client group in the perdag so it can compare its cookie with
the memdag's main branch's cookie.  Retrieving the BaseSnapshot commit currently requires walking the
full chain of local commits (retrieving them from idb).  While offline the chain of local commits continually grows
and so this can become problematically slow.

Solution
=======
Add a baseSnapshotHash to LocalMetaDD31 to allow for quick retrieval of the base snapshot commit.