feat: Allow push/pull to return ClientGroupNotFound
We now look at the result of push.

If the push/pull request responds with a JSON message `{error:
'ClientGroupNotFound'}` then we abort the pull/push and we mark the
client group as disabled.

A disabled client group will not be used for push/pull anymore.

A disabled client group cannot be reused when creating a new client but
we can fork from it.

We also store the client group disabled state in memory on the
replicache instance because reading that state out of the perdag is slow.

Towards #165