feat: Notify clients about new client groups (#415)
When a new client group is created we notify all the other Replicache
clients with the same name. These then "fire" onUpdateNeeded which
by default reloads the page.

A new client group is created when there isn't a compatible client group
in the perdag (indexeddb). When this happens the clients in different
client groups cannot sync when offline. They can only sync through the
server.

See:
https://www.notion.so/replicache/DD-3-1-e42489fc2e6b4340a01c7fa0de353a30#1b0ff4c4034846febda91af009dfd84a
and
https://www.notion.so/replicache/Version-incompatibilities-528f68f02f9c43f2904cdfef1e35e9c3

Based on #285

Towards #165