feat!: Remove createIndex and dropIndex (#464)
* fix!: Remove createIndex and dropIndex

BREAKING CHANGE: This removes the createIndex and dropIndex methods from
Replicache

This also removes the same methods from db.Write

There is some fallout in this PR where I started renaming types. Sorry.
The rule for now is that we should name things as FooDD31 and FooSDD and
we should export the DD31 version in src/mod.ts. We should remove the
doSomething methods that just delegate to the DD31 and SDD version based
on the compile time constant. This is so that we are explicit about things.

I deleted some dead code but I stopped since it was not related to this
PR. Must not clean things up!

Towards #165