feat: Introduce a cookie type (#484)
With DD31 we need cookies to be ordered. A cookie is either a string or
a number or an object with an `order` property. For the genesis commit
the value `null` is used for the cookie.

When comparing cookies where one cookies is a string and the other is a
number we compare the string representation of the number.

Closes #474
Towards #165