feat: Allow all CORS request, handle authentication and authorization via headers (#215)
Problem
=======
DD31 needs to be able to send pull requests from the client to the CF worker for mutation recovery.  This will pretty much always be a cross origin request.  

Solution
=======
Allow CORS for all request from all origins.  

Authentication and authorization are enforced via auth headers:
1. Sec-WebSocket-Protocol in the case of web socket connection request
2. x-reflect-auth-api-key for API requests
3. Authorization header for the soon to be introduced pull request for dd31 mutation recover

None of these rely on cookies, and we *must not* use cookies for auth.  Doing so with this allow all CORS policy will result in a CSRF vulnerability.