feat(DD31,reflect): Add a delay in refresh to favor pull/poke winning (#503)
Pull/poke and refresh are racing to see who gets to update the memdag
(the one with the newer basesnapshot cookie wins). pull/poke updates are
preferable to refresh udpates, so delay refresh slightly to make pull/poke
the winner except when pull/pokes are slow. This is especially important
for pokes, as refresh winning will result in the next poke's cookie not
matching necessitating a disconnect/reconnect.

This artificially inflates the refresh benchmarks, but by a fixed ammount.