chore: include clientID in all log messages.
problem: clientID was not included in early log messages because
it is an async field of Replicache but Reflect constructor is sync.

solution: make LogContext Promise<LogContext> and await it at call
sites.

Also removed GapTracker as it wasn't used in a long time and
fixed some usages of 'req' log context.