fix: lastServerAckdMutationIDs undefined (#506)
Issue:
When a client's lastServerAckdMutationIDs is undefined mutation recovery does not update it correctly due to a bad comparison of undefined to a number. lastServerAckdMutationIDs are updated by persist and mutation recovery, persist updates it to match the last mutation ids of the base snapshot being persisted so a client's lastServerAckdMutationIDs entry will be undefined until a snapshot containing a lastMutationId for the client is persisted.

Fix:
Update mutation recovery logic to treat undefined lastServerAckdMutationIDs values as zero so that they are updated correctly.