chore: Close with ErrorKind (#298)
Instead of sending an ErrorMessage followed by a close this now send the ErrorKind as the code in close.

This changes the enum from a string valued enum to a number valued enum,

WebSocket close allows passing a custom error code down that needs to be between 4000 and 4999. The ErrorKind enum therefore starts at 4000.

The WS close also allows passing a reason. However, the reason is limited to 123 utf-8 encoded bytes. For now that seems ok but if we need to pass more information in the error message we will have to revisit this again.

reflect pr is here: https://github.com/rocicorp/reflect/pull/81