feat: add a state metric to report last connect error
This metric will be reported if there was a connect error during
the reporting period and reflect still hasn't connected.

Also _disconnect() now takes a reason which we use in this change
and will use in a future metric that tracks the reason why short-lived
connections close.

We no longer throw when we get an error message on the connection
because many of these messages are expected, and they are fatal to
the connection. Instead we now disconnect.