feat: Connection work (#89)
This change moves a lot of the management of the connection into a "run
loop". The run loop replaces the watchdog.

A lot of the logic is still kept in the callbacks on the WebSocket.

The run loop deals with 2 states:
- Disconnected:
  - If we need to reauthenticate we do that before starting to connect.
  - Now it tries to connect
  - If connect fails it will reject and we enter the error path.
- Connected:
  - Here we wait for one of three cases:
    - We get disconnected
    - We get a message from the server
    - We get a timeout
      - Now it is time to do a PING.
  - If we get an error message or an error is thrown we increment an
    error count. The error count is used to determine how long to sleep
    after being disconnected before trying to reconnect.

The run does not have a state for Connecting. We do not leave the
disconnected state until we are connected.

This handles timeouts in Connect and Ping.

The reauth API is using Replicache API in this PR but it will be changed
in a followup PR.

Towards https://github.com/rocicorp/reflect-server/issues/294
Fixes https://github.com/rocicorp/reflect-server/issues/324
Fixes https://github.com/rocicorp/reflect-server/issues/325