feat: Refinements to onOnlineChange callback (#92)
This adds a getter for `online`.

We only call `onOnlineChange` when the online state changes. This is a
slight change from before because we assume we are online when starting
and the first connect will not call `onOnlineChange` since we are
already online. The old code always called `onOnlineChange` once at
startup.

The logic for calling `onOnlineChange` changed slightly. We do not
consider Reflect to be offline just because the websocket got closed. We
only consider it offline if the websocket was closed with an error or if
we fail to reconnect.

Fixes https://github.com/rocicorp/reflect-server/issues/333