feat: Add experimental Create KV store (#525)
This adds a new option called `experimentalCreateKVStore` which can be
used to create a KV store. This is useful for testing and for creating
custom stores.

The old `experimentalKVStore` option is still available but marked as deprecated.
It is broken due to IDB being used for the database of databases even when the
`experimentalKVStore` option is used.

This adds a `MemStore` class which is an in memory implementation of the
KV store interface. It takes a `name` parameter which is used to share
the data between named instances. This has similar semantics to the
`IDBStore`.

The `MemStore` is exported as `ExperimentalMemKVStore` for now.

Towards #434
Towards #483