fix: Relax isDeepFrozen check (#528)
If we expect people to be able to create custom Store implementations,
we need a way for them to be able to create deeply frozen objects.

However, previously, we were not doing a deep recursive check for deeply
frozen objects. Instead we tagged the objects (in a WeakSet) as we were
creating them, and then checked that the WeakSet contained the object
when we were checking if it was deeply frozen.

With this change we do the deep recursive check if the object is not in
the WeakSet, allowing us to support custom Store implementations.