feat: Check for version mismatch in the protocol (#353)
This checks the version in connect. We use a new URL for connect that matches `/api/sync/:version/connect`. If the version is not `v0` we return a `ErrorKind.VersionNotSupported` error.

For the old URL, `/connect`, we treat this as version 0 as well.

Fixes https://github.com/rocicorp/reflect-server/issues/331