refactor: move pull for mutation recovery to websocket messages (#20)
This gets rid of the recently introduced pull http endpoint.  
I decided to go forward with this as I think it simplifies things overall.  Once again all communication between the reflect client and server is over the web socket.  It eliminates the auth logic in the auth do for authing the pull endpoint.  It also simplifies versioning of the sync protocol (just the socket connect endpoint needs to be versioned).   Realizing I needed to add more versioning logic for the pull endpoint is what motivated me to go ahead and make this change.

It slightly complicates the reflect client in order to correctly route the response messages.

This is still not the ideal factoring for mutation recovery over websockets.  The main issue is mutation recovery pushes the mutations, and then immediately pulls.  With turn buffering the pushes mutations are not likely to have been applied by the time of the pull, meaning we will do a duplicated mutation recovery in the future.  This could be improved by:
1. updating mutation recovery to do a pull check, then push, then pull again.
2. or even better, have the server remember that client X has pushed mutations for client group id Y stored in indexdb name Z, and when it applies the mutations, send a message to X, with the updated last mutation ids and cookie, along with the client group id Y and indexdb name Z so that X can update the correct persisted state.