feat: Improve error messages for valita
We now have logic for generating error messages for valita errors. It is
questionable if it is really worth it.

All places now import shared/valita.js which has a slightly different
type for Type which does not include parse etc. This is because
we want to ensure we get our custom error messages.

Towards #216