Fix client-side caching
A naive approach of just changing to if env == client doesn't work, because the mutators need to run everywhere. This approach will result in work being done in multiple places and passing the entire cache as a mutator arg, but will no longer perform CPU-heavy work in the mutator path.