Update rebucketing
If a user goes offline, then the main room changes, then comes back online, we don't want the paint to abruptly change, so we just leave them in the old room.
Note that technically this means it's possible that we could go over the limit on that room, in the case that:
1. A user signs on
2. A user goes offline
3. >50 people join the room
4. that user comes back online

However, this room will only ever shrink, since when 50 people joined, the default room changed. So now nobody will ever be added to this room again.

I think this generally is the best tradeoff for UX vs limiting room size, but it's worth noting that this makes our room cap fairly soft during rushes.