fix: Initial fast-forward poke not sent until first mutation (#435)
This fix is based on the existing design for DisconnectHandler, and aims to make it easy to implement https://github.com/rocicorp/mono/issues/175.

Instead of adding connections to the stored `connectedclients` in handleConnection, we sync the ClientMap connections to the stored `connectedclients` when we process a turn.  A turn is run if there are pending mutations or there is a difference between the ClientMap and the stored `connectedclients`. 

In a follow up we can update `processDisconnects` to be `processConnectionChanges` and have it call both `DisconnectHandler` and a new `ConnectHandler`.