feat(60fps): dynamic buffer rework (#459)
Problem
=======
Existing implementation usesobserved variance in poke delivery to size the buffer.  This was giving poor results (buffer would often get adjusted down to something too small, and then take many adjustments to get up to a sufficient size).  While the buffer is meant to deal with variance, its an indirect relationship and I think this was just getting buffer size wrong.

Solution
======
Use a more direct measure for sizing the buffer.  For each frame, record the minimum buffer size that would have allowed the pokes in the frame to playback on time, this value is `bufferNeededMs` for that frame.  Every 10 seconds (and at least 200 frames), we adjust the buffer size up if more than 1% of frames were missed, or down if less than 0.5% of frames were missed.   We adjust it to the the 99.5 percentile of the recorded `bufferNeededMs` for that time frame.  

Also adds some additional perf debug fields to poke for debugging buffering behavior.