fix!: change how client heads work to fix ChunkNotFound (#465)
Instead of a headHash and tempRefreshHash, we now have: refreshHashes and persistHash.  

**refreshHashes**
refreshHashes is a set of hashes that replaces headHash and tempRefreshHash.  This change addresses a problem that occurs if the second perdag write tx in refresh fails, failing to update headHash and cleanup tempRefreshHash.  
refreshHashes contains:
1. The hash of the last commit this client refreshed from its client group (this is the commit it bootstrapped from until it completes its first refresh).
2. One or more hashes that were added to retain chunks of a commit while it was being refreshed into this client's memdag. (This can be one or more because refresh's cleanup step is a separate transaction and can fail).
Upon refresh completing and successfully running its clean up step, this set will contain a single hash: the hash of the last commit this client refreshed.

**persistHash**
Is new and most directly resolves the observed ChunkNotFound issue.  persistHash is the hash of the last snapshot commit persisted by this client to this client's client group, or null if has never persisted a snapshot.

In persist, when we write the memdag's basesnapshot to the perdag we tell the memdag via call to chunksPersisted that it can move these chunks from its `_memOnlyChunks` (where chunks can not be evicted) to its `_sourceChunksCache` (where chunks can be evicted).  However, the only thing retaining these chunks in the perdag is the ClientGroup's head hash, which can be updated by another client.   This new `persistHash` will retain these chunks in the case that another client updates the ClientGroup's head hash.  

Fixes #434 #30 

BREAKING CHANGE: Updates format version from 5 to 6 because the structure of persisted Client(s) in the ClientMap is changed.