Fix weird mouseover behavior when pieces are overlapped.
Before when two pieces overlapped, you could not mouse back and
forth between them and get hover effect on both. This was due to
a temporary workaround inthe code, but the workaround was only
necessary due to incorrect/reversed z-index logic. With the
z-index logic fixed the workaround no longer necessary and
hover effects work correctly.