Move back button under pieces.
It would be even better if pieces couldn't overlap the back button,
but this is an easy quick thing. It's better for a piece to block
the back button because you can just move it out of the way. And
exiting the puzzle if you mistap is non-fatal because you can just
go back in.