fix: puzzle did not display when navigating back to site in ios. (#142)
This was a bfcache thing. When we press the back button, we don't
get a new load cycle. However pagehide had already unhooked the
periodic alive ping for orchestration.

I'm not sure why we were unhooking that on page hide. There's no
need to stop the timer since (a) it doesn't send a ping if the
tab is not visible and (b) background tabs will naturally have
their timers rate limited by the browser anyway.

Tested navigation to and from and tab switching in ios/safari,
desktop safari, and desktop chrome -- everything seems to work.