fix!: Fix lost DataDog logs issue due to reused ReflectServerOptions (#531)
Problem
======
Sometimes logs stop showing up in DataDog). It's due to these two behaviors of CloudFlare DOs.
If a setTimeout callback is pending when a DO is killed that setTimeout never fires. 
Sometimes a DO is restarted in the same isolate (I have verified this happens when you call wrangler tail or wrangler publish with no code changes, I could see them doing this under other circumstances as well). That is it has the same execution context/global memory as a previous instance of the DO.
As of this [change](https://github.com/rocicorp/mono/commit/c882ec2da6719927e564f7249187e1c7bd716682) we cache ReflectServerOptions globaly for a given isolate.   That means the DataDogLogSink is a singleton for a given isolate.  If a DO is ever killed while the [setTimeout](https://github.com/rocicorp/mono/blob/main/packages/datadog/src/datadog-log-sink.ts#L48) for flushing the DataDogLogSink is pending, we end up with a DataDogLogSink with a non-zero this._timerID, but no pending setTimeout to clear it.  Going forward no logs will be flushed for this isolate.

Solution
======
Given CloudFlare's behavior, the safest thing is to ensure that everything used by a DO is scoped to that DO, i.e. is created when the DO is created and is stored in its instance vars. API wise, this means we should get rid of the non-function ReflectServerOptions option, and call the function to create the options everytime a DO is constructed.  

However, this still doesn't completely protect us from this CloudFlare behavior. Our customer's could think they are being efficient by constructing a DataDogLogSink once, storing it in a global variable, and always returning this singleton instance from their ReflectServerOptions function.  I've left a comment warning against doing this.

BREAKING CHANGE: createReflectServer no longer accepts non-function options.