fix: make current schema version 0 add more logging/dont await flush in ensureStorageSchemaMigrated (#533)
The down migrate code (which would have never run), would have written version 1.  Instead lets make it 0, since we are not writing 1 in the forward migrate direction.

Hopefully this will fix, or at least enable us to debug, the "A call to blockConcurrencyWhile() in a Durable Object waited for too long." errors Monday experienced when deploying v0.24.0.